Reorganize the tests for the operations;
PWPA-2002
