+++ src/alfasim_score/converter/alfacase/_tests/test_build_production_alfacase/test_create_alfacase_natural_flow_production.alfacase
