+++ src/alfasim_score/converter/alfacase/_tests/test_convert_base_alfacase/test_create_alfacase_natural_flow.alfacase
